Refactor into `AudioUtils`
